requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
numpy==1.24.3
scipy==1.11.2
